fixed first bug

+++ lib/provider/user_provider.dart
@@ -27,22 +27,39 @@ class UserProvider extends ChangeNotifier {
   SubscriptionModel? _myPlan;
 
   User? get currentUser => _currentUser;
+
   String? get userId => _userId;
+
   String? get ownerId => _ownerId;
+
   String? get budgetId => _budgetId;
+
   String? get profileImageUrl => _profileImageUrl;
+
   UserModel? get myInfo => _myInfo;
+
   bool get imageExists => _imageExists;
+
   bool get isLoggedIn => _userId != null;
+
   bool get justSignedIn => _justSignedIn;
+
   List<BudgetModel>? get budgets => _budgets;
+
   List<BudgetModel>? get myBudgets => _myBudgets;
+
   List<String>? get permissionBudgets => _permissionBudgets;
+
   List<String>? get sharedOwnerIds => _sharedOwnerIds;
+
   List<String>? get sharedUserIds => _sharedUserIds;
+
   List<UserModel>? get sharedUsers => _sharedUsers;
+
   List<UserModel>? get mySharedUsers => _mySharedUsers;
+
   List<UserModel>? get sharedOwnerUsers => _sharedOwnerUsers;
+
   SubscriptionModel? get myPlan => _myPlan;
 
   Future<void> initializeUserProvider() async {
@@ -50,16 +67,12 @@ class UserProvider extends ChangeNotifier {
     _userId = _currentUser?.id;
 
     if (_currentUser != null) {
-      final response = await Supabase.instance.client
-          .from('users')
-          .select('*')
-          .eq('id', _userId!)
-          .single();
+      final response =
+          await Supabase.instance.client.from('users').select('*').eq('id', _userId!).single();
       _myInfo = UserModel.fromJson(response);
       _myPlan = await loadUserSubscription(_userId!);
-      _profileImageUrl = Supabase.instance.client.storage
-          .from('avatars')
-          .getPublicUrl('$_userId/profile.png');
+      _profileImageUrl =
+          Supabase.instance.client.storage.from('avatars').getPublicUrl('$_userId/profile.png');
       await loadSharedUsers();
       await loadBudgets();
     } else {
@@ -119,7 +132,7 @@ class UserProvider extends ChangeNotifier {
       _budgetId = lastBudgetId;
     } else {
       final mainBudget = _budgets!.firstWhere(
-            (b) => b.isMain == true,
+        (b) => b.isMain == true,
         orElse: () => _budgets!.first,
       );
       _budgetId = mainBudget.id;
@@ -163,8 +176,7 @@ class UserProvider extends ChangeNotifier {
     if (_userId != null) {
       await Supabase.instance.client
           .from('users')
-          .update({'last_budget_id': _budgetId})
-          .eq('id', _userId!);
+          .update({'last_budget_id': _budgetId}).eq('id', _userId!);
     }
   }
 
@@ -202,7 +214,7 @@ class UserProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-    Future<void> _validationBudgetId() async {
+  Future<void> _validationBudgetId() async {
     final response = await Supabase.instance.client
         .from('users')
         .select('last_budget_id')
@@ -226,10 +238,8 @@ class UserProvider extends ChangeNotifier {
     final sharedOwnerGroup = responseGetOwner.map(SharedUserModel.fromJson).toList();
     _sharedOwnerIds = sharedOwnerGroup.map((e) => e.ownerId!).toList();
 
-    final responseSharedOwnerUser = await Supabase.instance.client
-        .from('users')
-        .select('*')
-        .inFilter('id', _sharedOwnerIds!);
+    final responseSharedOwnerUser =
+        await Supabase.instance.client.from('users').select('*').inFilter('id', _sharedOwnerIds!);
     _sharedOwnerUsers = responseSharedOwnerUser.map(UserModel.fromJson).toList();
 
     // 현재 내가 선택하고 있는 ownerId를 공유받고 있는 userId
@@ -241,10 +251,8 @@ class UserProvider extends ChangeNotifier {
     _sharedUserIds = sharedUserGroup.map((e) => e.userId!).toList();
 
     // 현재 ownerID 그룹에 포함되어 있는 user들으 의 정보 가져옴
-    final userResponse = await Supabase.instance.client
-        .from('users')
-        .select('*')
-        .inFilter('id', _sharedUserIds!);
+    final userResponse =
+        await Supabase.instance.client.from('users').select('*').inFilter('id', _sharedUserIds!);
     _sharedUsers = userResponse.map(UserModel.fromJson).toList();
 
     final responseGetSharedUser = await Supabase.instance.client
@@ -254,32 +262,24 @@ class UserProvider extends ChangeNotifier {
     final mySharedUserGroup = responseGetSharedUser.map(SharedUserModel.fromJson).toList();
     final mySharedUserIds = mySharedUserGroup.map((e) => e.userId!).toList();
 
-    final mySharedUserResponse = await Supabase.instance.client
-        .from('users')
-        .select('*')
-        .inFilter('id', mySharedUserIds);
+    final mySharedUserResponse =
+        await Supabase.instance.client.from('users').select('*').inFilter('id', mySharedUserIds);
     _mySharedUsers = mySharedUserResponse.map(UserModel.fromJson).toList();
     notifyListeners();
   }
 
   Future<void> loadBudgets() async {
-    final response = await Supabase.instance.client
-        .from('budgets')
-        .select('id')
-        .eq('owner_id', _ownerId!);
+    final response =
+        await Supabase.instance.client.from('budgets').select('id').eq('owner_id', _ownerId!);
     final allBudgetIds = response.map(BudgetModel.fromJson).toList().map((e) => e.id!).toList();
 
     final permissionResponse = await Supabase.instance.client
         .from('budget_permissions')
         .select('budget_id')
         .eq('user_id', _userId!);
-    final permittedBudgetIds = permissionResponse
-        .map((e) => e['budget_id'] as String)
-        .toList();
+    final permittedBudgetIds = permissionResponse.map((e) => e['budget_id'] as String).toList();
 
-    _permissionBudgets = allBudgetIds
-        .where((id) => permittedBudgetIds.contains(id))
-        .toList();
+    _permissionBudgets = allBudgetIds.where((id) => permittedBudgetIds.contains(id)).toList();
 
     final finalBudgetResponse = await Supabase.instance.client
         .from('budgets')
@@ -287,10 +287,8 @@ class UserProvider extends ChangeNotifier {
         .inFilter('id', _permissionBudgets!);
     _budgets = finalBudgetResponse.map(BudgetModel.fromJson).toList();
 
-    final myBudgetResponse = await Supabase.instance.client
-        .from('budgets')
-        .select('*')
-        .eq('owner_id', _userId!);
+    final myBudgetResponse =
+        await Supabase.instance.client.from('budgets').select('*').eq('owner_id', _userId!);
     _myBudgets = myBudgetResponse.map(BudgetModel.fromJson).toList();
     notifyListeners();
   }
@@ -300,7 +298,7 @@ class UserProvider extends ChangeNotifier {
       final response = await http.head(Uri.parse(_profileImageUrl!));
       _imageExists = response.statusCode == 200;
     } catch (e) {
-      _imageExists =  false;
+      _imageExists = false;
     }
     notifyListeners();
   }
@@ -320,10 +318,12 @@ class UserProvider extends ChangeNotifier {
     });
 
     if (fullUserIds.isNotEmpty) {
-      final rows = fullUserIds.map((userId) => {
-        'budget_id': newBudgetId,
-        'user_id': userId,
-      }).toList();
+      final rows = fullUserIds
+          .map((userId) => {
+                'budget_id': newBudgetId,
+                'user_id': userId,
+              })
+          .toList();
 
       await supabase.from('budget_permissions').insert(rows);
     }
@@ -347,10 +347,12 @@ class UserProvider extends ChangeNotifier {
 
     // 3. 새로운 권한 삽입
     if (fullUserIds.isNotEmpty) {
-      final rows = fullUserIds.map((userId) => {
-        'budget_id': budgetId,
-        'user_id': userId,
-      }).toList();
+      final rows = fullUserIds
+          .map((userId) => {
+                'budget_id': budgetId,
+                'user_id': userId,
+              })
+          .toList();
 
       await supabase.from('budget_permissions').insert(rows);
     }
@@ -363,10 +365,8 @@ class UserProvider extends ChangeNotifier {
     final supabase = Supabase.instance.client;
 
     // 1. 이 budget을 참조하는 user 모두 가져오기
-    final referencingUsers = await supabase
-        .from('users')
-        .select('id')
-        .eq('last_budget_id', budgetId);
+    final referencingUsers =
+        await supabase.from('users').select('id').eq('last_budget_id', budgetId);
 
     for (final user in referencingUsers) {
       final userId = user['id'];
@@ -382,26 +382,17 @@ class UserProvider extends ChangeNotifier {
       final mainBudgetId = mainBudget?['id'];
 
       // 3. users 테이블 업데이트
-      await supabase
-          .from('users')
-          .update({
+      await supabase.from('users').update({
         'last_owner_id': userId,
         'last_budget_id': mainBudgetId,
-      })
-          .eq('id', userId);
+      }).eq('id', userId);
     }
 
     // 1. 권한 정보 먼저 삭제
-    await supabase
-        .from('budget_permissions')
-        .delete()
-        .eq('budget_id', budgetId);
+    await supabase.from('budget_permissions').delete().eq('budget_id', budgetId);
 
     // 2. 예산 자체 삭제
-    await supabase
-        .from('budgets')
-        .delete()
-        .eq('id', budgetId);
+    await supabase.from('budgets').delete().eq('id', budgetId);
 
     // 3. 현재 선택 중인 예산이면 교체
     if (_budgetId == budgetId) {
@@ -451,9 +442,7 @@ class UserProvider extends ChangeNotifier {
         userId: response['user_id'],
         planName: response['plan_name'],
         startDate: DateTime.parse(response['start_date']),
-        endDate: response['end_date'] != null
-            ? DateTime.parse(response['end_date'])
-            : null,
+        endDate: response['end_date'] != null ? DateTime.parse(response['end_date']) : null,
         isActive: response['is_active'],
         adsEnabled: planInfo['ads_enabled'],
         maxBudgets: planInfo['max_budgets'],
